refactor orderdetails for 2 apps
